requirements.txt
accelerate==0.27.2
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
diffusers==0.21.2
executing==2.0.1
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
huggingface-hub==0.17.2
icecream==2.1.3
idna==3.6
importlib_metadata==7.0.2
importlib_resources==6.1.3
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
psutil==5.9.8
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.3.3
six==1.16.0
sympy==1.12
tokenizers==0.13.3
torch==2.2.1
tqdm==4.66.2
transformers==4.33.2
triton==2.2.0
typing_extensions==4.10.0
urllib3==2.2.1
zipp==3.17.0
